requirements.txt
apache-libcloud==1.5.0
argcomplete==1.9.2
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
cycler==0.10.0
decorator==4.1.2
idna==2.5
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
matplotlib==2.0.2
nbformat==4.4.0
numpy==1.13.0
pandas==0.20.2
paramiko==2.2.1
plotly==2.1.0
pyasn1==0.2.3
pycparser==2.17
PyNaCl==1.1.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
readline==6.2.4.1
requests==2.18.4
six==1.10.0
traitlets==4.3.2
urllib3==1.22
